method Line.normalize now expects double and returns int to avoid casting argument to int in every call

+++ src/main/java/ru/vsu/cs/cg/g8_1/kilchenko_v_p/lagrangepolynomial/Line.java
@@ -5,14 +5,14 @@ import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.paint.Color;
 
 public class Line {
-    private static int normalize(int n) {
-        return Math.max(Math.min(n, 3000), -1);
+    private static int normalize(double n) {
+        return (int) Math.max(Math.min(n, 3000), -1);
     }
     public static void lineFromTo(Point2D from, Point2D to, Color startColor, Color endColor, GraphicsContext graphicsContext) {
-        int x0 = normalize((int) from.getX());
-        int y0 = normalize((int) from.getY());
-        int x1 = normalize((int) to.getX());
-        int y1 = normalize((int) to.getY());
+        int x0 = normalize(from.getX());
+        int y0 = normalize(from.getY());
+        int x1 = normalize(to.getX());
+        int y1 = normalize(to.getY());
 
         int dx = Math.abs(x1 - x0);
         int dy = Math.abs(y1 - y0);
